django_room_booking_api/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
pytz==2023.3
Pillow==10.1.0 